[WINDOWS] New releases bring new challenges to windows build :-)
