+++ Source/plugins/JSONRPC.cpp
@@ -17,29 +17,31 @@
  * limitations under the License.
  */
 
-#pragma once
-
-#include "Module.h"
+#include "JSONRPC.h"
 
 namespace WPEFramework {
+
 namespace PluginHost {
 
+    JSONRPC::JSONRPC()
+        : _adminLock()
+        , _handlers()
+        , _service(nullptr)
+    {
+        std::vector<uint8_t> versions = { 1 };
 
-    class EXTERNAL IFactories {
-    public:
-        virtual ~IFactories();
+        _handlers.emplace_back([&](const uint32_t id, const string& designator, const string& data) { Notify(id, designator, data); }, versions);
+    }
 
-        static void Assign(IFactories* instance);
-        static IFactories& Instance();
-        static bool IsAvailable();
+    JSONRPC::JSONRPC(const std::vector<uint8_t> versions)
+        : _adminLock()
+        , _handlers()
+        , _service(nullptr)
+    {
+        _handlers.emplace_back([&](const uint32_t id, const string& designator, const string& data) { Notify(id, designator, data); }, versions);
+    }
+    /* virtual */ JSONRPC::~JSONRPC()
+    {
+    }
 
-    public:
-        virtual Core::ProxyType<Web::Request> Request() = 0;
-        virtual Core::ProxyType<Web::Response> Response() = 0;
-        virtual Core::ProxyType<Web::FileBody> FileBody() = 0;
-        virtual Core::ProxyType<Web::JSONBodyType<Core::JSONRPC::Message>> JSONRPC() = 0;
-    };
-
-    typedef DEPRECATED Core::WorkerPool WorkerPool;
-    typedef DEPRECATED IFactories Factories;
 } }
