+++ Source/plugins/plugins.vcxproj
@@ -20,13 +20,14 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Channel.cpp" />
-    <ClCompile Include="JSONRPC.cpp" />
     <ClCompile Include="MetaData.cpp" />
     <ClCompile Include="Module.cpp" />
     <ClCompile Include="Service.cpp" />
     <ClCompile Include="Shell.cpp" />
+    <ClCompile Include="JSONRPC.cpp" />
     <ClCompile Include="StateControl.cpp" />
     <ClCompile Include="SubSystem.cpp" />
+    <ClCompile Include="System.cpp" />
     <ClCompile Include="VirtualInput.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -44,6 +45,7 @@
     <ClInclude Include="plugins.h" />
     <ClInclude Include="Request.h" />
     <ClInclude Include="Service.h" />
+    <ClInclude Include="System.h" />
     <ClInclude Include="VirtualInput.h" />
   </ItemGroup>
   <ItemGroup>
